requirements.txt
# Core libraries
pandas>=1.3.0
numpy>=1.20.0
requests>=2.25.0
beautifulsoup4>=4.9.0
tqdm>=4.60.0
Pillow>=8.2.0

# PDF processing
PyPDF2>=2.0.0
pdf2image>=1.16.0
pytesseract>=0.3.8

# AI/ML
google-generativeai>=0.1.0

# Utilities
python-dateutil>=2.8.1